requirements.txt
anthropic>=0.40.0
groq>=0.9.0
google-genai>=1.0.0
python-dotenv>=1.0.0
python-frontmatter>=1.1.0
streamlit>=1.30.0
